Moved test_utils to a higher folder
