+++ tests/info/test_CURRENT_ENVIRONMENT.py
@@ -4,7 +4,7 @@ import imp
 
 import SwissKnife
 import SwissKnife.info
-import tests.info as test_utils
+import tests.test_utils as test_utils
 
 class Test_CURRENT_ENVIRONMENT(unittest.TestCase):
